Server/requirements.txt
acme==0.21.1
asn1crypto==0.24.0
backports-abc==0.5
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.6.0
bs4==0.0.1
certbot==0.21.1
certifi==2017.7.27.1
cffi==1.11.4
chardet==3.0.4
Cheetah==2.4.4
cloud-init==0.7.6
ConfigArgParse==0.12.0
configobj==4.7.2
cryptography==2.1.4
decorator==3.4.0
enum34==1.1.6
funcsigs==1.0.2
future==0.16.0
futures==3.2.0
icalendar==4.0.3
idna==2.6
iniparse==0.4
ipaddress==1.0.19
IPy==0.75
javapackages==1.0.0
jedi==0.11.0
Jinja2==2.9.6
josepy==1.0.1
jsonpatch==1.16
jsonpointer==1.10
kitchen==1.1.1
lxml==3.2.1
Markdown==2.6.8
MarkupSafe==1.0
mock==2.0.0
oauth==1.0.1
parsedatetime==2.4
parso==0.1.0
pbr==3.1.1
perf==0.1
PIL==1.1.7
Pillow==5.2.0
policycoreutils-default-encoding==0.1
prettytable==0.7.2
pycparser==2.18
pycurl==7.19.0
pygobject==3.22.0
pygpgme==0.3
pyliblzma==0.5.3
pyOpenSSL==17.5.0
pyRFC3339==1.0
pytesseract==0.2.2
python-dateutil==2.7.5
python-linux-procfs==0.4.9
pytz==2017.3
pyudev==0.15
pyxattr==0.5.1
PyYAML==3.12
requests==2.18.4
schedutils==0.4
seobject==0.1
sepolicy==1.1
singledispatch==3.4.0.3
six==1.11.0
slip==0.4.0
slip.dbus==0.4.0
timeout-decorator==0.4.1
tornado==4.5.2
urlgrabber==3.10
urllib3==1.22
virtualenv==1.10.1
web.py==0.39
yum-metadata-parser==1.1.4
zope.component==4.4.1
zope.event==4.3.0
zope.interface==4.4.3
